norm, some remaining (25 lines, +4 fun args)

+++ new_minishell/srcs/utils/ft_is_alphanum.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_is_alphanum.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gurousta <gurousta@student.42.fr>          +#+  +:+       +#+        */
+/*   By: vicalvez <vicalvez@student.42nice.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/23 19:59:34 by gurousta          #+#    #+#             */
-/*   Updated: 2024/03/23 20:05:18 by gurousta         ###   ########.fr       */
+/*   Updated: 2024/04/04 16:50:33 by vicalvez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 int	ft_is_alphanum(char c)
 {
-	size_t	index;
-	
+	int	index;
+
 	index = 0;
 	while (index < 10)
 	{
